crud operations done for entites but still relations crud operations under development 9:03am commit

+++ Models/Products.cs
@@ -7,8 +7,8 @@ namespace store1.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float wholesalePrice { get; set; }
-        public float price { get; set; }
+        public double wholesalePrice { get; set; }
+        public double price { get; set; }
         public List<Customer> Customers { get; set; }
     }
 }
